drop shell script fusion
