--- clear.sh
@@ -1,4 +1,0 @@
-rm -rf Dia/**/migrations
-rm -rf Dia/**/migrations/*
-rm -rf Dia/**/migrations/00*
-rm -rf Dia/**/migrations/__pycache__
